requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==22.1.0
boto3==1.21.41
botocore==1.24.46
certifi==2021.10.8
charset-normalizer==2.1.0
click==8.1.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.1
fastapi_auth_middleware==1.0.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pydantic==1.9.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
python-jose==3.3.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
rsa==4.8
s3transfer==0.5.2
sendgrid==6.9.7
six==1.16.0
sniffio==1.2.0
starkbank-ecdsa==2.2.0
starlette==0.17.1
tomli==2.0.1
twilio==7.16.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.2